requirements.txt
annotated-types==0.6.0
anyio==4.2.0
APScheduler==3.10.4
atproto==0.0.42
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.3
dnspython==2.6.1
h11==0.14.0
httpcore==1.0.3
httpx==0.25.2
idna==3.6
libipld==1.1.0
pillow==10.2.0
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
typing_extensions==4.9.0
tzlocal==5.2
urllib3==2.2.1
websockets==12.0